+++ src/widgets/conditions/condition-record.css
@@ -1,0 +1,38 @@
+:global(.omrs-breakpoint-gt-phone) .conditionSummary {
+  display: flex;
+}
+
+:global(.omrs-breakpoint-lt-tablet) .conditionSummary {
+  padding: 0.25rem 0.5rem 0 0.5rem;
+}
+
+.conditionCard {
+  margin: 0rem 1rem;
+}
+
+.conditionTable {
+  width: 100%;
+  margin-top: 1rem;
+  margin-bottom: 1rem;
+}
+
+.conditionTable thead th {
+  color: var(--omrs-color-ink-medium-contrast);
+  text-align: left;
+}
+
+.conditionTable thead {
+  color: var(--omrs-color-ink-medium-contrast);
+}
+
+.conditionTable tbody td:nth-child(1) {
+  font-family: "Work Sans";
+}
+
+.conditionTable thead td {
+  margin-top: 0.625rem;
+}
+
+.conditionTable thead td:nth-child(1) {
+  width: 25%;
+}
