+++ src/widgets/conditions/conditions-detailed-summary.css
@@ -1,38 +1,121 @@
 :global(.omrs-breakpoint-gt-phone) .conditionSummary {
   display: flex;
+  justify-content: center;
 }
 
 :global(.omrs-breakpoint-lt-tablet) .conditionSummary {
   padding: 0.25rem 0.5rem 0 0.5rem;
 }
 
-.conditionCard {
-  margin: 0rem 1rem;
-}
-
 .conditionTable {
   width: 100%;
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-}
-
-.conditionTable thead th {
-  color: var(--omrs-color-ink-medium-contrast);
-  text-align: left;
 }
 
 .conditionTable thead {
   color: var(--omrs-color-ink-medium-contrast);
-}
-
-.conditionTable tbody td:nth-child(1) {
-  font-family: "Work Sans";
+  border-bottom: 0.0625rem solid var(--omrs-color-bg-low-contrast);
 }
 
 .conditionTable thead td {
-  margin-top: 0.625rem;
+  padding: 1rem 0.25rem 1rem 1rem;
 }
 
-.conditionTable thead td:nth-child(1) {
-  width: 25%;
+:global(.omrs-breakpoint-lt-tablet) .conditionTable thead td {
+  padding: 0.25rem 0.125rem 0.5rem 0.5rem;
+}
+
+.conditionTable thead td:nth-child(2) {
+  text-align: right;
+  padding-right: 0rem;
+}
+
+.conditionTable thead td:nth-child(3) {
+  text-align: left;
+  padding-right: 2.925rem;
+}
+
+.conditionTable tbody td:nth-child(1) {
+  width: 15%;
+}
+
+.conditionTable tbody td:nth-child(2) {
+  width: 19%;
+  text-align: right;
+  font-family: "Work Sans";
+}
+
+.conditionTable tbody td:nth-child(3) {
+  width: 15%;
+  text-align: right;
+  font-family: "Work Sans";
+}
+
+.conditionTable tbody td:nth-child(4) {
+  width: 15%;
+  padding-right: 1.125rem;
+  color: var(--omrs-color-bg-low-contrast);
+  text-align: right;
+  font-family: "Work Sans";
+}
+
+.conditionTable tbody tr:nth-child(3n) {
+  border-bottom: 0.0625rem solid var(--omrs-color-bg-low-contrast);
+}
+
+.conditionTable tbody {
+  padding: 2rem;
+  border-bottom: 0.0625rem solid var(--omrs-color-bg-low-contrast);
+}
+
+:global(.omrs-breakpoint-lt-tablet) .conditionTable tbody {
+  padding: 1rem;
+}
+
+.conditionTable tr {
+  padding: 0.813rem 0 1rem 0.813rem;
+  border-bottom: 0.0625rem solid var(--omrs-color-bg-low-contrast);
+}
+
+.conditionTable td {
+  padding: 0.75rem 0 0.75rem 1rem;
+}
+
+.centerItems {
+  display: flex;
+  align-items: center;
+}
+
+.alignRight {
+  justify-content: flex-end;
+}
+
+.alignLeft {
+  justify-content: flex-start;
+}
+
+.alignRight span {
+  margin-right: 1rem;
+}
+
+.conditionFooter {
+  text-align: center;
+  font-family: "Work Sans";
+  color: var(--omrs-color-ink-medium-contrast);
+}
+
+.conditionMargin {
+  margin: 1rem 1rem;
+}
+
+.conditionMargin p {
+  padding-bottom: 1rem;
+}
+
+tr.inactive {
+  background: var(--omrs-color-bg-medium-contrast);
+  color: var(--omrs-color-ink-medium-contrast);
+}
+
+tr.active {
+  background: var(--omrs-color-bg-high-contrast);
 }
