+++ src/widgets/level-two-routes.component.tsx
@@ -7,11 +7,11 @@ import ProgramsSummary from "../widgets/programs/programs-summary.component";
 import ProgramsDetailedSummary from "../widgets/programs/programs-detailed-summary.component";
 import AllergyForm from "../widgets/allergies/allergy-form.component";
 import VitalsForm from "../widgets/vitals/vitals-form.component";
-import ConditionsSummary from "../widgets/conditions/conditions-summary.component";
+import ConditionsDetailedSummary from "./conditions/conditions-detailed-summary.component";
 import MedicationLevelTwo from "../widgets/medications/medication-level-two.component";
 import MedicationCardLevelThree from "./../widgets/medications/medication-level-three/medication-level-three.component";
 import AllergyCardLevelThree from "../widgets/allergies/allergy-card-level-three.component";
-import ConditionsDetailedSummary from "../widgets/conditions/conditions-detailed-summary.component";
+import ConditionRecord from "./conditions/condition-record.component";
 import MedicationOrderBasket from "./medications/medication-order-basket.component";
 import MedicationOrder from "./medications/medication-order.component";
 import HeightAndWeightDetailedSummary from "../widgets/heightandweight/heightandweight-detailed-summary.component";
@@ -34,12 +34,12 @@ export const levelTwoRoutes: PatientChartRoute[] = [
   },
   {
     url: "/patient/:patientUuid/chart/conditions",
-    component: ConditionsSummary,
+    component: ConditionsDetailedSummary,
     name: "Conditions"
   },
   {
     url: "/patient/:patientUuid/chart/conditions/:conditionUuid",
-    component: ConditionsDetailedSummary,
+    component: ConditionRecord,
     name: "Conditions"
   },
   {
